// sineedge_front/src/app/settings/model/branch.ts
export class Branch {
    // ID: number;
    // CityID: string;
    // Name: string;
    branch_id: any; 
    fk_city_id: any; 
    name: any; 
    createdon: any; 
    fk_createdby: any; 
    updatedon: any; 
    fk_updatedby: any; 
    isactive: any;
}
export class States {
    state_id:  any;  
    name:  any;    
    isactive: any;
}
export class City {
    city_id:  any; 
    name:  any; 
    isactive: any;
}